+++ packages/amazon-bedrock/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @ai-sdk/amazon-bedrock
 
+## 3.0.93
+
+### Patch Changes
+
+- Updated dependencies [a27a978]
+  - @ai-sdk/provider-utils@3.0.23
+  - @ai-sdk/anthropic@2.0.74
+
+## 3.0.92
+
+### Patch Changes
+
+- Updated dependencies [42bc840]
+  - @ai-sdk/anthropic@2.0.73
+
+## 3.0.91
+
+### Patch Changes
+
+- Updated dependencies [98bd0e2]
+  - @ai-sdk/anthropic@2.0.72
+
+## 3.0.90
+
+### Patch Changes
+
+- 0967b6c: fix(provider/amazon-bedrock): transform bedrock/anthropic error responses to anthropic format
+
 ## 3.0.89
 
 ### Patch Changes