+++ packages/rsc/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,79 @@
 
 ### Patch Changes
 
+- ai@5.0.173
+
+## 0.0.1
+
+### Patch Changes
+
+- ai@5.0.172
+
+## 0.0.1
+
+### Patch Changes
+
+- ai@5.0.171
+
+## 0.0.1
+
+### Patch Changes
+
+- ai@5.0.170
+
+## 0.0.1
+
+### Patch Changes
+
+- ai@5.0.169
+
+## 0.0.1
+
+### Patch Changes
+
+- ai@5.0.168
+
+## 0.0.1
+
+### Patch Changes
+
+- ai@5.0.167
+
+## 0.0.1
+
+### Patch Changes
+
+- ai@5.0.166
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [a27a978]
+  - ai@5.0.165
+
+## 0.0.1
+
+### Patch Changes
+
+- ai@5.0.164
+
+## 0.0.1
+
+### Patch Changes
+
+- ai@5.0.163
+
+## 0.0.1
+
+### Patch Changes
+
+- ai@5.0.162
+
+## 0.0.1
+
+### Patch Changes
+
 - ai@5.0.161
 
 ## 0.0.1