+++ packages/vercel/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @ai-sdk/vercel
 
+## 1.0.36
+
+### Patch Changes
+
+- Updated dependencies [a27a978]
+  - @ai-sdk/provider-utils@3.0.23
+  - @ai-sdk/openai-compatible@1.0.35
+
 ## 1.0.35
 
 ### Patch Changes